+++ backend/simulator/src/main/java/com/chaoxu/simulator/optimizer/L2Obj.java
@@ -3,9 +3,9 @@ package com.chaoxu.simulator.optimizer;
 import java.util.Map;
 
 public class L2Obj extends Objective {
-    public double value(Map<String, Double> waitingTime) {
+    public double value(Map<String, Integer> waitingTime) {
         double ret = 0;
-        for (double x : waitingTime.values()) {
+        for (int x : waitingTime.values()) {
             ret += x * x;
         }
         return ret / waitingTime.size();
